refactor(axios folder): remove lib folder, adding axios.ts in util
